move back atomic result
https://coveord.atlassian.net/browse/KIT-1861
