+++ packages/atomic/src/components/common/atomic-result-placeholder/atomic-result-placeholder.pcss
@@ -1,4 +1,4 @@
-@import '../atomic-result/atomic-result.pcss';
+@import '../result/result.pcss';
 
 .result-root {
   &.display-grid {
