+++ packages/atomic/src/components/common/result/result-cell-desktop.pcss
