+++ packages/atomic/src/components/common/result/result-cell-mobile.pcss
