+++ packages/atomic/src/components/common/result/result-row-desktop.pcss
