+++ packages/atomic/src/components/common/result/result-row-mobile.pcss
