+++ packages/atomic/src/components/common/result/result-with-sections.pcss
