+++ packages/atomic/src/components/common/result/result.pcss
